Fix group handling
Add size() method for PusherGroup.  Also, remember to add a new pusher
from a new group to that group after creating the group (doh!).  Also,
when DeviceRegistry is asked for a particular group, don't just return
a null pointer if there are no pushers in that group;  instead create
and return an empty list, so that strips.size() can return 0 rather
than throwing a null pointer exception.