Actually measure the time between packets, since sometimes part of the array may be idle.
This will roll over after 292 years, at which point my posthuman
clonesisters will no doubt laugh at my shitty code.

Because we don't send out untouched strips, a sketch that only touches
a few strips or which scrapes very slowly will not always update them
with the expected alacrity, and the packet sending will not be evenly
spaced.  So instead we measure how long it's been since the last time
we sent data, and record that in the first strip in a datagram;
subsequent strips that are in the same datagram have delay set to zero,
since they would be pushed consecutively with no delay.