Made locking on strips consistent.
Locking on strips was inconsistent.

Removed the Semaphore for striplock and replaced with an object to
synchronize on. Locking is now performed on all access, mutation and
iteration of strips by holding the stripLock.

Added lazy creation of strips in setAntiLog method. There was a race
condition that could intermittently cause a NullPointerException.

Removed copying of the strips array in copyHeader. It looks like the
PixelManager reference in the copied Strips will be incorrect, so it is
safer to re-create the pixels on demand.