Support for multicasting to pushers
If a pusher announces itself on a multicast IP, we consider it to be a
multicast target.  The first device on that multicast group will be a
Primary, and all other devices will be a Stooge.  Multicast Stooges do
not return strips, and their CardThread does not push data.  Unicast
Stooges continue to do their thing as before.