Add a mechanism for strip age
The getPushedAt() method in Strip allows you to determine how recently
a strip was pushed.  Since PixelPusher is asynchronous, it might sit in
the pipeline for some time, and this allows applications to determine
how long.

A value of 0 indicates that the strip has not yet been pushed since it
was last updated;  otherwise, the value is in the same timebase as
System.nanoTime().